+++ github/commitstatus.go
@@ -11,7 +11,6 @@ import (
 	"io"
 	"net/http"
 	"path"
-	"slices"
 	"strconv"
 	"strings"
 	"time"
@@ -34,18 +33,10 @@ func (e *StatusError) Error() string {
 const API = "https://api.github.com"
 
 type Target struct {
+	// Server is the GitHub API server.
+	// Currently, hardcoded to https://api.github.com
 	Server string
-	// MaxAttempts is the maximum number of attempts when retrying an HTTP request to
-	// GitHub, no matter the reason (rate limited or transient error).
-	MaxAttempts int
-	// WaitTransient is the wait time before the next attempt when encountering a
-	// transient error from GitHub.
-	WaitTransient time.Duration
-	// MaxSleepRateLimited is the maximum sleep time (over all attempts) when rate
-	// limited from GitHub.
-	MaxSleepRateLimited time.Duration
-	// Jitter is added to the sleep duration to prevent creating a Thundering Herd.
-	Jitter time.Duration
+	Retry  Retry
 }
 
 // CommitStatus is a wrapper to the GitHub API to set the commit status for a specific
@@ -54,13 +45,11 @@ type Target struct {
 // - NewCommitStatus
 // - https://docs.github.com/en/rest/commits/statuses
 type CommitStatus struct {
-	target  Target
+	target  *Target
 	token   string
 	owner   string
 	repo    string
 	context string
-
-	sleepFn func(d time.Duration) // Overridable by tests.
 
 	log hclog.Logger
 }
@@ -76,14 +65,13 @@ type CommitStatus struct {
 //
 // See also:
 // - https://docs.github.com/en/rest/commits/statuses
-func NewCommitStatus(target Target, token, owner, repo, context string, log hclog.Logger) CommitStatus {
+func NewCommitStatus(target *Target, token, owner, repo, context string, log hclog.Logger) CommitStatus {
 	return CommitStatus{
 		target:  target,
 		token:   token,
 		owner:   owner,
 		repo:    repo,
 		context: context,
-		sleepFn: time.Sleep,
 		log:     log,
 	}
 }
@@ -133,45 +121,23 @@ func (cs CommitStatus) Add(sha, state, targetURL, description string) error {
 	req.Header.Set("Accept", "application/vnd.github.v3+json")
 	req.Header.Set("Content-Type", "application/json")
 
-	var response httpResponse
-
-	for attempt := 1; ; attempt++ {
-		s.log.Info("GitHub HTTP request", "attempt", attempt, "max", s.target.MaxAttempts)
-		response, err = httpRequestDo(req)
-		if err != nil {
-			return err
-		}
-		if attempt == s.target.MaxAttempts {
-			break
-		}
-		retry, timeToSleep, reason := checkForRetry(response, s.target.WaitTransient,
-			s.target.MaxSleepRateLimited, s.target.Jitter)
-		if !retry {
-			break
-		}
-		s.log.Info("Sleeping for", "duration", timeToSleep, "reason", reason)
-		s.sleepFn(timeToSleep)
+	response, retryErr := cs.target.Retry.Do(cs.log, func() (HttpResponse, error) {
+		return httpRequestDo(req)
+	})
+	if retryErr != nil {
+		return fmt.Errorf("commit status: %s", retryErr)
 	}
 
 	return cs.checkStatus(response, state, sha, url)
 }
 
-type httpResponse struct {
-	statusCode         int
-	body               string
-	oauthInfo          string
-	rateLimitRemaining int
-	rateLimitReset     time.Time
-	date               time.Time
-}
-
-func httpRequestDo(req *http.Request) (httpResponse, error) {
-	var response httpResponse
+func httpRequestDo(req *http.Request) (HttpResponse, error) {
+	var response HttpResponse
 	// By default, there is no timeout, so the call could hang forever.
 	client := &http.Client{Timeout: time.Second * 30}
 	resp, err := client.Do(req)
 	if err != nil {
-		return httpResponse{}, fmt.Errorf("http client Do: %w", err)
+		return HttpResponse{}, fmt.Errorf("http client Do: %w", err)
 	}
 	defer resp.Body.Close()
 
@@ -216,14 +182,14 @@ func httpRequestDo(req *http.Request) (httpResponse, error) {
 	// https://datatracker.ietf.org/doc/html/rfc2616#section-14.18
 	date, err := time.Parse(time.RFC1123, resp.Header.Get("Date"))
 	if err != nil {
-		return httpResponse{}, fmt.Errorf("failed to parse the date header: %s", err)
+		return HttpResponse{}, fmt.Errorf("failed to parse the date header: %s", err)
 	}
 	response.date = date
 
 	return response, nil
 }
 
-func (cs CommitStatus) checkStatus(resp httpResponse, state, sha, url string) error {
+func (cs CommitStatus) checkStatus(resp HttpResponse, state, sha, url string) error {
 	var hint string
 
 	switch resp.statusCode {
@@ -242,7 +208,7 @@ func (cs CommitStatus) checkStatus(resp httpResponse, state, sha, url string) er
 		hint = "Either wrong credentials or PAT expired (check your email for expiration notice)"
 	case http.StatusForbidden:
 		if resp.rateLimitRemaining == 0 {
-			hint = fmt.Sprintf("Rate limited but the wait time to reset would be longer than %v (MaxSleepRateLimited)", cs.target.MaxSleepRateLimited)
+			hint = fmt.Sprintf("Rate limited but the wait time to reset would be longer than %v (MaxSleepRateLimited)", cs.target.Retry.MaxSleepRateLimited)
 		} else {
 			hint = "none"
 		}
@@ -264,62 +230,4 @@ OAuth: %s`,
 			url,
 			resp.oauthInfo),
 	}
-}
-
-// checkForRetry determines if the HTTP request should be retried after a sleep.
-// If yes, checkForRetry returns true, the sleep duration and a reason.
-// If no, checkForRetry returns false and a reason.
-//
-// To take a decision, use only the boolean value. Do not use the duration nor the reason.
-//
-// It considers two different reasons for a retry:
-//  1. The request encountered a GitHub-specific rate limit.
-//     In this case, it considers parameters maxSleepTime and jitter.
-//  2. The HTTP status code is in a retryable subset of the 5xx status codes.
-//     In this case, it returns the same as the input parameter waitTime.
-func checkForRetry(res httpResponse, waitTime, maxSleepTime, jitter time.Duration,
-) (bool, time.Duration, string) {
-	retryableStatusCodes := []int{
-		http.StatusInternalServerError, // 500
-		http.StatusBadGateway,          // 502
-		http.StatusServiceUnavailable,  // 503
-		http.StatusGatewayTimeout,      // 504
-	}
-
-	// Are we rate limited ?
-	// If the request exceeds the rate limit, the response will have status 403 Forbidden
-	// and the x-ratelimit-remaining header will be 0
-	// https://docs.github.com/en/rest/overview/resources-in-the-rest-api?apiVersion=2022-11-28#exceeding-the-rate-limit
-	if res.statusCode == http.StatusForbidden && res.rateLimitRemaining == 0 {
-		// Calculate the sleep time based solely on the server clock. This is unaffected
-		// by the inevitable clock drift between server and client.
-		sleepTime := res.rateLimitReset.Sub(res.date)
-		// Be robust to possible races in the GitHub backend. This avoids failing too early.
-		if sleepTime < 0 {
-			sleepTime = 0
-		}
-		// Be a good netizen by adding some jitter to the time we sleep.
-		sleepTime += jitter
-
-		if sleepTime > maxSleepTime {
-			return false, 0, "rate limited, sleepTime > maxSleepTime, should not retry"
-		}
-		return true, sleepTime, "rate limited, should retry"
-	}
-
-	// Do we have a retryable HTTP status code ?
-	if slices.Contains(retryableStatusCodes, res.statusCode) {
-		return true, waitTime, http.StatusText(res.statusCode)
-	}
-
-	// The status code could be 200 OK or any other error we did not process before.
-	// In any case, there is nothing to sleep, return 0 and let the caller take a
-	// decision.
-	return false, 0, "no retryable reasons"
-}
-
-// SetSleepFn overrides time.Sleep, used when retrying an HTTP request, with sleepFn.
-// WARNING Use only in tests.
-func (s *CommitStatus) SetSleepFn(sleepFn func(d time.Duration)) {
-	s.sleepFn = sleepFn
 }