+++ github/commitstatus_private_test.go
@@ -11,7 +11,7 @@ import (
 func TestCheckForRetry(t *testing.T) {
 	type testCase struct {
 		name       string
-		res        httpResponse
+		res        HttpResponse
 		waitTime   time.Duration
 		jitter     time.Duration
 		wantRetry  bool
@@ -34,19 +34,19 @@ func TestCheckForRetry(t *testing.T) {
 	testCases := []testCase{
 		{
 			name:       "status OK: do not retry",
-			res:        httpResponse{statusCode: http.StatusOK},
+			res:        HttpResponse{statusCode: http.StatusOK},
 			wantRetry:  false,
 			wantReason: "no retryable reasons",
 		},
 		{
 			name:       "non retryable status code: do not retry",
-			res:        httpResponse{statusCode: http.StatusTeapot},
+			res:        HttpResponse{statusCode: http.StatusTeapot},
 			wantReason: "no retryable reasons",
 			wantRetry:  false,
 		},
 		{
 			name:       "retryable status code: retry, sleep==waitTime",
-			res:        httpResponse{statusCode: http.StatusInternalServerError},
+			res:        HttpResponse{statusCode: http.StatusInternalServerError},
 			waitTime:   42 * time.Second,
 			wantRetry:  true,
 			wantSleep:  42 * time.Second,
@@ -54,7 +54,7 @@ func TestCheckForRetry(t *testing.T) {
 		},
 		{
 			name: "rate limited, would sleep too long: do not retry",
-			res: httpResponse{
+			res: HttpResponse{
 				statusCode:     http.StatusForbidden,
 				date:           serverDate,
 				rateLimitReset: serverDate.Add(30 * time.Minute),
@@ -64,7 +64,7 @@ func TestCheckForRetry(t *testing.T) {
 		},
 		{
 			name: "rate limited, do retry, sleep adding also the jitter",
-			res: httpResponse{
+			res: HttpResponse{
 				statusCode:     http.StatusForbidden,
 				date:           serverDate,
 				rateLimitReset: serverDate.Add(5 * time.Minute),
@@ -79,7 +79,7 @@ func TestCheckForRetry(t *testing.T) {
 			name: "same server date and rateLimitReset, zero jitter, repro of Pix4D/cogito#124",
 			// Same server date and rateLimitReset.
 			// This can be explained by a benign race in the backend.
-			res: httpResponse{
+			res: HttpResponse{
 				statusCode:     http.StatusForbidden,
 				date:           serverDate,
 				rateLimitReset: serverDate,
@@ -93,7 +93,7 @@ func TestCheckForRetry(t *testing.T) {
 			name: "robust against server date after rateLimitReset",
 			// Server date slightly after rateLimitReset.
 			// This can be explained by a benign race in the backend.
-			res: httpResponse{
+			res: HttpResponse{
 				statusCode:     http.StatusForbidden,
 				date:           serverDate,
 				rateLimitReset: serverDate.Add(-2 * time.Second),