+++ CHANGELOG.md
@@ -5,7 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## UNRELEASED
+## UNRELEASED [v0.11.0] - 2023-08-09
+
+### Added 
+
+- Retry GitHub API HTTP request when rate limited or when transient server error:
+  - New generic `retry` package, usable with any API (not GitHub specific).
+  - New Retry adapters for the GitHub API in the `github` package, usable for any GitHub API, not only the Commit Statuses API.
+  - Easily integrable also by tools using the go-github package, although not dependent upon it.
+
+### Removed
+
+- Prior code to retry GitHub API HTTP request when rate limited or when transient server error.
 
 ### Changed