+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## UNRELEASED [v0.11.0] - 2023-09-XX
+## [v0.11.0] - 2023-09-22
 
 ### Added 
 
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
-- Prior code to retry GitHub API HTTP request when rate limited or when transient server error.
+- Previous code to retry GitHub API HTTP request when rate limited or when transient server error.
 
 ### Changed